Nueva relacion & Hub

+++ WillyNet.Charlas.Core.Domain/Entities/UserApp.cs
@@ -11,7 +11,7 @@ namespace WillyNet.Charlas.Core.Domain.Entities
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public short Dni { get; set; }
+        public int Dni { get; set; }
         public ICollection<Asistencia> Asistencias { get; set; }
         public ICollection<Control> Controls { get; set; }
     }
